+++ scripts/presentation/common.mjs
@@ -1,16 +1,55 @@
 import { SETTINGS } from "../settings.mjs"
 
-export function getAllItems(type) {
-    return game.packs.filter(p => p.metadata.type === "Item")
-        .map(
-            p => p.index.filter(item => item.type === type).map(item => item.uuid)
-        ).flat()
+/**
+ * Retrieves all item UUIDs of a specified type from a single pack.
+ * Applies special filtering to handle feats
+ *
+ * @param {string} type - The type of items to retrieve.
+ * @param {strimg} pack - The pack to retrieve items from 
+ * @returns {Promise<string[]>} An array of item UUIDs of the specified type.
+ */
+async function getItemsFromPack(type, pack) {
+    if (type === 'feat') {
+        return await pack.getIndex(
+            {fields: new Set(["uuid", "system.type"])}
+        )
+        .then(
+            index =>  index.filter(d => d.system.type?.value === type)
+                .map(d => d.uuid)
+        )
+    }
+    return await pack.index.filter(d => d.type === type).map(item => item.uuid)
 }
 
+/**
+ * Retrieves all item UUIDs of a specified type from the game packs.
+ *
+ * @param {string} type - The type of items to retrieve.
+ * @returns {string[]} An array of item UUIDs of the specified type.
+ */
+export async function getAllItems(type) {
+    return (await Promise.all(game.packs.filter(p => p.metadata.type === "Item")
+        .map(
+            async (p) => await getItemsFromPack(type, p)
+        ))).flat()
+}
+
+
+/**
+ * Creates a new journal entry with the specified title, sheet, journal type, and folder.
+ * If a journal entry with the same name and type already exists in the specified folder, it will be deleted first.
+ *
+ * @param {string} title - The title of the journal entry.
+ * @param {string} sheet - The sheet class to be used for the journal entry.
+ * @param {string} journalType - The type of the journal entry.
+ * @param {Folder|null} [folder=null] - The folder in which to create the journal entry. If null, the journal entry will be created in the root folder.
+ * @returns {Promise<JournalEntry|null>} - The newly created journal entry, or null if creation failed.
+ */
 export async function makeJournal(title, sheet, journalType, folder=null) {
 
+
     //Delete existing journal if present
-    const existingJournal = game.journal.filter(p => p.folder === folder 
+    const existingJournal = game.journal.filter(p => p.folder?.uuid === folder?.uuid 
         && p.name === SETTINGS[journalType].label 
         && p.flags.dcm.journalType === journalType
     )
@@ -42,8 +81,16 @@ export async function makeJournal(title, sheet, journalType, folder=null) {
     return newJournal;
 }
 
-export async function makePage(title, content, level=null, sort=null) {
-//Make a basic page with title and content
+/**
+ * Creates a basic page object with a title and content.
+ *
+ * @param {string} title - The title of the page.
+ * @param {string} content - The content of the page.
+ * @param {number|null} [level=null] - The level of the title (optional).
+ * @param {number|null} [sort=null] - The sort order of the page (optional).
+ * @returns {Object} The page object.
+ */
+export function makePage(title, content, level=null, sort=null) {
     return {
         _id: foundry.utils.randomID(),
         name: title,
@@ -53,7 +100,39 @@ export async function makePage(title, content, level=null, sort=null) {
     };
 }
 
+/**
+ * Creates a spell page object.
+ *
+ * @param {string} title - The title of the spell page.
+ * @param {string[]} spells - An array of spell UUIDs.
+ * @param {number} level - The spell level.
+ * @param {number|null} [sort=null] - The sort order.
+ * @returns {Object} A promise that resolves to the spell page object.
+ */
+export function makeSpellPage(title, spells, level, sort=null) {
+    return {
+        _id: foundry.utils.randomID(),
+        name: title,
+        type: "spells",
+        sort: sort,
+        title: {level :level},
+        system: {
+            grouping: "level",
+            identifier: "classIdentifier",
+            type: "class",
+            spells: spells
+        }
+    }
+}
 
+
+/**
+ * Adds pages to a journal entry.
+ *
+ * @param {JournalEntry} journal - The journal entry to which pages will be added.
+ * @param {Array<Object>} pages - An array of page objects to be added to the journal entry.
+ * @returns {Promise<Array<JournalEntryPage>>} A promise that resolves to an array of the newly created journal entry pages.
+ */
 export async function addPages(journal, pages) {
     const embeddedPages = []
     for (const p of pages) {
@@ -63,8 +142,14 @@ export async function addPages(journal, pages) {
 }
 
 
+/**
+ * Generates an HTML table from the provided headers and rows.
+ *
+ * @param {string[]} headers - An array of strings representing the table headers.
+ * @param {string[][]} rows - A 2D array of strings representing the table rows.
+ * @returns {string} - A string containing the HTML representation of the table.
+ */
 export function htmlTable(headers, rows) {
-
     const toElement = (element, h) => `<${element}>${h}</${element}>`
 
     const htmlRows = []