+++ scripts/presentation/create-classbook.mjs
@@ -1,5 +1,5 @@
 import {getSetting, SETTINGS} from '../settings.mjs';
-import { getAllItems, makeJournal, makePage, addPages } from './common.mjs';
+import { getAllItems, makeJournal, makePage, addPages, makeSpellPage } from './common.mjs';
 
 
 /**
@@ -84,18 +84,11 @@ function makeCombinedSpellList(className, classIdentifier) {
             [...spells.uuids]
 
     // Create new spell list page data
-    return {
-        name: `${className} Spell List`,
-        type: "spells",
-        sort: null,
-        title: {level :2},
-        system: {
-            grouping: "level",
-            identifier: "classIdentifier",
-            type: "class",
-            spells: filteredSpells
-        }
-    }    
+    return makeSpellPage(
+        `${className} Spell List`,
+        filteredSpells,
+        2
+    )
 }
 
 /**
@@ -144,12 +137,12 @@ export async function createClassbook(folder, useExistingPages, sheet) {
     // Get classes based on settings or get all available classes
     const classes = getSetting(SETTINGS.class.enabled)
         ? getSetting(SETTINGS.class.content) 
-        : getAllItems("class");
+        : await getAllItems("class");
     
     // Get subclasses based on settings or get all available subclasses
     const subclasses = getSetting(SETTINGS.subclass.enabled)
         ? getSetting(SETTINGS.subclass.content)
-        : getAllItems("subclass");
+        : await getAllItems("subclass");
     
     // Build index of existing pages if requested
     const existingPageIndex = useExistingPages