+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+### 1.4.2
+New Features:
+- Improvement of how duplicate filtering works and addition of an inverse filter to find unique
+- Button to collapse/expand tables of items for easier working with large sets of items
+Bug Fixes
+- "Make Spells journal" option in Create Player Handbook now actually creates a journal
+- Fixed QuickInsert integration bug which caused it not to function until changes we made
+- Feat handbook no longer contains non-feat features if when not being managed by DCM
+
 ### 1.4.1
 - Fix for Quick Insert integration broken by latest QI update
 - Retain settings when generating player handbooks